Refined instructions and tests

+++ src/main/object/oriented/java/figures/Figure.java
@@ -4,6 +4,6 @@ import object.oriented.java.SystemException;
 
 public interface Figure {
 
-	public Integer getArea() throws SystemException;
+	public Double getArea() throws SystemException;
 
 }
