+++ src/main/object/oriented/java/figures/Square.java
@@ -2,11 +2,11 @@ package object.oriented.java.figures;
 
 public class Square implements Figure {
 
-	public int base;
-	public Integer area = null;
+	public Integer base;
+	public Double area = null;
 
-	public Integer getArea() {
-		area = base * base;
+	public Double getArea() {
+		area = Double.valueOf(base * base);
 		return area;
 	}
